Replace reverse-engineered m-n expansions with ...
... a version based on the original paper _Adaptive Precision
Floating-Point Arithmetic and Fast Robust Geometric Predicates_. The
paper describes an algorithm `EXPANSION-SUM` and provides pseudocode.
`predicates.c` uses this pattern and a similar pattern for variants of
"EXPANSION-DIFF" (e.g. `Two_Two_Diff()`). The `DIFF` variants of the
algorithm don't exist in any paper, so we still had to rely on a bit of
reverse-engineering for those rules.

Ultimately though, this means we can tune performance to use
`EXPANSION-SUM`, `LINEAR-EXPANSION-SUM`, or `FAST-EXPANSION-SUM`, the
lattter two with and without Zero Elimination, at various stages of each
predicate pipeline, to determine which is better on modern hardware.